More plugins compile

+++ plugins/src/math.oscillators/module_plugin_maths_oscillators_inside_range.h
@@ -131,7 +131,7 @@ public:
     {
       is_in_range = true;
       every_beat->set(1.0f);
-      float random_number = rand() % 100;
+      float random_number = (float)(rand() % 100);
       if (random_number > randomness->get())
       {
         random_beat->set(0.0f);
