+++ plugins/src/mesh.generators/module_mesh_plane_uv_distort.h
@@ -177,13 +177,13 @@ public:
         int z_index8192 = (int)round( 8192.0f * iz1 );
         // calculate skewed uv coord
         vsx_tex_coord2f uv_skew;
-        uv_skew.s = ix1 + x_shape[x_index8192] * x_shape_multiplier_f;
-        uv_skew.t = iz1 + z_shape[z_index8192] * z_shape_multiplier_f;
+        uv_skew.s = (float)(ix1 + x_shape[x_index8192] * x_shape_multiplier_f);
+        uv_skew.t = (float)(iz1 + z_shape[z_index8192] * z_shape_multiplier_f);
 
         mesh->data->vertices[vi] =  vsx_vector3<>
                                     (
-                                      2.0f * (ix1 - 0.5f),
-                                      2.0f * (iz1 - 0.5f),
+                                      2.0f * ((float)ix1 - 0.5f),
+                                      2.0f * ((float)iz1 - 0.5f),
                                       0
                                     );
